requirements.txt
flask
pymupdf
pytesseract
pillow
python-dotenv
requests